requirement.txt
asgiref==3.4.1
attrs==21.4.0
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2.10
django-braces==1.15.0
django-crispy-forms==1.13.0
django-gravatar2==1.4.4
django-model-utils==4.2.0
django-render-partial==0.4
django-star-ratings==0.9.2
django-widget-tweaks==1.4.9
idna==3.3
isodate==0.6.1
lxml==4.7.1
Pillow==8.4.0
platformdirs==2.4.1
python-decouple==3.5
pytz==2021.3
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
typing_extensions==4.0.1
urllib3==1.26.8
zeep==4.1.0
